creating-deleting data by using token-userId

+++ controllers/expenses.js
@@ -2,6 +2,7 @@
 const Expense = require('../models/expense');
 const User = require('../models/user');
 
+
 exports.addExpense = async (req, res, next) => {
 
     try {
@@ -13,7 +14,8 @@ exports.addExpense = async (req, res, next) => {
         res.status(401).json({message: 'please enter all the fields'});
     }
 
-    const data = await Expense.create({amount, description, category});
+    const data = await Expense.create({amount, description, category, userId: req.user.id});
+
     res.status(200).json({expDetails: data});
 
 }catch(err) {console.log(err)};
@@ -23,9 +25,11 @@ exports.addExpense = async (req, res, next) => {
 
 exports.getExpenses = async (req, res, next) => {
     try {
-        const data = await Expense.findAll();
+        // const data = await Expense.findAll();
+       const data = await Expense.findAll({where: {userId: req.user.id}}); //getting expense of user who is logged in
         res.status(200).json({allDetails: data})
     }catch(err) {
+        console.log(err);
         res.status(500).json({error:err})
     };
    
@@ -34,7 +38,12 @@ exports.getExpenses = async (req, res, next) => {
 exports.deleteExpense = async (req, res, next) => {
     try {
         const expId = req.params.id;
-        await Expense.destroy({where: {id: expId}})
-        res.status(200).json({message:'user deleted successfully'});
+        // await Expense.destroy({where: {id: expId}})
+        const noOfRows = await Expense.destroy({where: {id: expId, userId: req.user.id}}) //user can only delete his data not others
+        if(noOfRows === 0) {
+            res.status(404).json({success: false, message: 'You cant delete expesnes of others'});
+        }
+    
+        return res.status(200).json({message:'user deleted successfully'});
     }catch(err) {console.log(err)};
 }
